requirements.txt
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==6.0.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
decorator==5.2.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-rq==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
gunicorn==23.0.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.2
packaging==25.0
pillow==11.3.0
platformdirs==4.3.7
proglog==0.1.12
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
redis==6.4.0
requests==2.32.4
rq==2.4.1
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.29.3
whitenoise==6.9.0
